Add files via upload
Created the first Windows Form that uses AForge library to implement Camera. Able to start and stop the recording, also save the photo.
Notes:
Before using(Windows):Search bar->Camera Privacy Settings->Turn on "Allow apps to access your camera";
The saved image is always in it's default resolution quality despite the resolution changes. Needed for later use when scanning the photo.
When saving the photo adding the format is not necessary.
To run the program, you need only the files in the debug folder.